requirements.txt
pydantic-ai==0.0.53
pydantic==2.11.3

# Core
langchain==0.3.23
langgraph==0.3.30
langchain-community==0.3.21
langchain-ollama==0.3.2
langchain-chroma==0.2.3

# Resume parsing
pdfplumber==0.10.2
pyresparser==1.0.6
spacy==3.7.4
python-docx==1.1.2
textract==1.6.5
docx2txt==0.9

#RAG
chromadb==1.0.4
six==1.17.0
